Refactor ApacheResponseConsumerDispatching
- extract nested classes the separate files
- avoid using coroutines internal API
- supply more comments to clarify logic
