+++ ktor-client/ktor-client-apache/jvm/src/io/ktor/client/engine/apache/ApacheResponseConsumerDispatching.kt
@@ -12,7 +12,6 @@ import org.apache.http.nio.*
 import org.apache.http.nio.protocol.*
 import org.apache.http.protocol.*
 import java.nio.*
-import java.util.concurrent.*
 import kotlin.coroutines.*
 
 internal class ApacheResponseConsumerDispatching(
@@ -20,8 +19,7 @@ internal class ApacheResponseConsumerDispatching(
     private val block: (HttpResponse, ByteReadChannel) -> Unit
 ) : HttpAsyncResponseConsumer<Unit>, CoroutineScope {
     private val interestController = InterestControllerHolder()
-
-    private val dispatcher = ReactorLoopDispatcher()
+    private val dispatcher = ReactorLoopDispatcher(interestController, 1)
 
     override val coroutineContext: CoroutineContext = callContext + dispatcher
 
@@ -52,42 +50,67 @@ internal class ApacheResponseConsumerDispatching(
 
     private val contentChannel: ByteReadChannel = job.channel
 
+    /**
+     * We hold job completion exception here instead of calling job.getCancellationException
+     */
+    private val jobCompletionCause = atomic<Throwable?>(null)
+
+    /**
+     * Wait for a [ContentDecoder] ready for decoding (probably empty but unprocessed yet).
+     */
     private suspend fun waitForDecoder(): ContentDecoder? = suspendCancellableCoroutine {
         decoderWaiter = it
     }
 
     init {
-        // the coroutine is dispatched but not yet started
-        // (since we control it's execution using the custom dispatcher)
-        // so we start it's execution here so it should suspend at decoder suspension
+        // The coroutine is already started and dispatched but not yet executed
+        // (since it may only run inside of [ReactorLoopDispatcher])
+        // So we start its execution here, and it should suspend at [waitForDecoder] invocation.
         processLoop(Result.failure(IllegalStateException("The coroutine shouldn't be suspended at this point yet.")))
 
         check(decoderWaiter != null) { "Writer coroutine should suspend until decoder available." }
+
+        job.invokeOnCompletion { cause ->
+            jobCompletionCause.value = cause
+        }
+
+        jobCompletionCause.value?.let {
+            // rethrow if the coroutine crashed immediately
+            throw it
+        }
     }
 
     override fun consumeContent(decoder: ContentDecoder, ioctrl: IOControl) {
         do {
+            // resume if suspended before (this is double-check from a previous loop iteration)
             interestController.resumeInputIfPossible()
+
+            // run dispatchers loop (the coroutine) sending the decoder to it
             processLoop(Result.success(decoder))
 
             when {
                 decoderWaiter != null -> {
-                    // if the coroutine is waiting for decoder
+                    // If the coroutine is waiting for a decoder again
                     // then we don't touch interest
                     // let's wait for the next consumeContent invocation
                 }
                 job.isActive -> {
-                    // the coroutine is running somewhere
+                    // The coroutine is in suspension somewhere else (possibly due to back-pressure).
                     interestController.suspendInput(ioctrl)
                     // no double check needed here since we do loop and resume input at the loop beginning
                 }
                 else -> {
-                    // job is cancelled or crashes so we simply discard all incoming bytes reported by Apache
+                    // The coroutine is cancelled or crashed, so we simply discard all incoming bytes provided by Apache
                     if (!decoder.isCompleted) {
                         decoder.discardAll()
                     }
                 }
             }
+
+            // We retry the loop if there were tasks enqueued.
+            // This can only happen if the coroutine has been resumed
+            // from another thread. For example, a channel reader released some space in the channel,
+            // thus, we may continue execution.
         } while (dispatcher.hasTasks())
     }
 
@@ -96,19 +119,17 @@ internal class ApacheResponseConsumerDispatching(
         processLoop(Result.failure(ex))
     }
 
-    @UseExperimental(InternalCoroutinesApi::class)
     override fun cancel(): Boolean {
         job.cancel()
-        processLoop(Result.failure(job.getCancellationException()))
+        processLoop(Result.failure(IllegalStateException("Job cancellation should do resume.")))
         return true
     }
 
     override fun close() {
     }
 
-    @UseExperimental(InternalCoroutinesApi::class)
     override fun getException(): Exception? {
-        return job.getCancellationException().cause as? Exception
+        return jobCompletionCause.value as? Exception
     }
 
     override fun getResult() {
@@ -117,6 +138,7 @@ internal class ApacheResponseConsumerDispatching(
     override fun isDone(): Boolean = job.isCompleted
 
     override fun responseCompleted(context: HttpContext) {
+        // Passing null should cause the coroutine to exit normally if not yet.
         processLoop(Result.success(null))
     }
 
@@ -124,6 +146,9 @@ internal class ApacheResponseConsumerDispatching(
         block(response, contentChannel)
     }
 
+    /**
+     * Process dispatchers event loop resuming decoder waiting with the specified [result].
+     */
     private fun processLoop(result: Result<ContentDecoder?>) {
         decoderWaiter?.let { continuation ->
             this.decoderWaiter = null
@@ -133,53 +158,14 @@ internal class ApacheResponseConsumerDispatching(
         dispatcher.processLoop()
     }
 
+    /**
+     * Discard all ready bytes in [ContentDecoder].
+     */
     private fun ContentDecoder.discardAll() {
         val buffer = ByteBuffer.allocate(8192)
         do {
             buffer.clear()
             if (read(buffer) <= 0) break
         } while (true)
-    }
-
-    /**
-     * Holder class to guard reference to [IOControl] so one couldn't access it improperly.
-     */
-    private class InterestControllerHolder {
-        /**
-         * Contains [IOControl] only when it is suspended. One should steal it first before requesting input again.
-         */
-        private val interestController = atomic<IOControl?>(null)
-
-        fun suspendInput(ioControl: IOControl) {
-            ioControl.suspendInput()
-            interestController.update { before ->
-                check(before == null || before === ioControl) { "IOControl is already published" }
-                ioControl
-            }
-        }
-
-        fun resumeInputIfPossible() {
-            interestController.getAndUpdate { before ->
-                if (before == null) return
-                null
-            }?.requestInput()
-        }
-    }
-
-    private inner class ReactorLoopDispatcher : CoroutineDispatcher() {
-        private val queue = ArrayBlockingQueue<Runnable>(2)
-
-        override fun dispatch(context: CoroutineContext, block: Runnable) {
-            queue.add(block)
-            interestController.resumeInputIfPossible()
-        }
-
-        fun processLoop() {
-            while (true) {
-                queue.poll()?.run() ?: break
-            }
-        }
-
-        fun hasTasks(): Boolean = queue.isNotEmpty()
     }
 }
